server/nlp_service/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
chromadb==0.5.0
nltk==3.6.3
torch>=1.9.0
transformers>=4.11.0
sentence-transformers==2.2.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.0.0 